[tinker] Avoid suicide after system theme was changed.
