+++ tinker-android/tinker-android-loader/src/main/java/com/tencent/tinker/loader/TinkerResourcePatcher.java
@@ -27,17 +27,13 @@ import android.content.pm.ApplicationInfo;
 import android.content.res.AssetManager;
 import android.content.res.Resources;
 import android.os.Build;
-import android.os.Handler;
-import android.os.Message;
 import android.util.ArrayMap;
 
 import com.tencent.tinker.loader.shareutil.ShareConstants;
 import com.tencent.tinker.loader.shareutil.SharePatchFileUtil;
 import com.tencent.tinker.loader.shareutil.ShareReflectUtil;
-import com.tencent.tinker.loader.shareutil.ShareTinkerInternals;
 import com.tencent.tinker.loader.shareutil.ShareTinkerLog;
 
-import java.io.File;
 import java.io.InputStream;
 import java.lang.ref.WeakReference;
 import java.lang.reflect.Constructor;
@@ -45,7 +41,6 @@ import java.lang.reflect.Field;
 import java.lang.reflect.Method;
 import java.util.Collection;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
 /**
@@ -195,6 +190,7 @@ class TinkerResourcePatcher {
      * @param externalResourceFile
      * @throws Throwable
      */
+    @SuppressWarnings({"ConstantConditions", "unchecked"})
     public static void monkeyPatchExistingResources(Context context, String externalResourceFile, boolean isReInject) throws Throwable {
         if (externalResourceFile == null) {
             return;
@@ -228,7 +224,7 @@ class TinkerResourcePatcher {
 
         if (isReInject) {
             ShareTinkerLog.i(TAG, "Re-injecting, skip rest logic.");
-            recordCurrentPatchedResModifiedTime(externalResourceFile);
+            PatchedResourcesInsuranceLogic.recordCurrentPatchedResModifiedTime(externalResourceFile);
             return;
         }
 
@@ -237,7 +233,7 @@ class TinkerResourcePatcher {
         if (((Integer) addAssetPathMethod.invoke(newAssetManager, externalResourceFile)) == 0) {
             throw new IllegalStateException("Could not create new AssetManager");
         }
-        recordCurrentPatchedResModifiedTime(externalResourceFile);
+        PatchedResourcesInsuranceLogic.recordCurrentPatchedResModifiedTime(externalResourceFile);
 
         // Add SharedLibraries to AssetManager for resolve system resources not found issue
         // This influence SharedLibrary Package ID
@@ -314,154 +310,9 @@ class TinkerResourcePatcher {
             throw new TinkerRuntimeException(ShareConstants.CHECK_RES_INSTALL_FAIL);
         }
 
-        installResourceInsuranceHacks(context, externalResourceFile);
-    }
-
-    private static void installResourceInsuranceHacks(Context context, String patchedResApkPath) {
-        try {
-            final Object activityThread = ShareReflectUtil.getActivityThread(context, null);
-            final Field mHField = ShareReflectUtil.findField(activityThread, "mH");
-            final Handler mH = (Handler) mHField.get(activityThread);
-            final Field mCallbackField = ShareReflectUtil.findField(Handler.class, "mCallback");
-            final Handler.Callback originCallback = (Handler.Callback) mCallbackField.get(mH);
-            if (!(originCallback instanceof ResourceInsuranceHandlerCallback)) {
-                final ResourceInsuranceHandlerCallback hackCallback = new ResourceInsuranceHandlerCallback(
-                        context, patchedResApkPath, originCallback, mH.getClass());
-                mCallbackField.set(mH, hackCallback);
-            } else {
-                ShareTinkerLog.w(TAG, "installResourceInsuranceHacks: already installed, skip rest logic.");
-            }
-        } catch (Throwable thr) {
-            ShareTinkerLog.printErrStackTrace(TAG, thr, "failed to install resource insurance hack.");
-        }
-    }
-
-    private static final class ResourceInsuranceHandlerCallback implements Handler.Callback {
-        private static final String LAUNCH_ACTIVITY_LIFECYCLE_ITEM_CLASSNAME = "android.app.servertransaction.LaunchActivityItem";
-
-        private final Context mContext;
-        private final String mPatchResApkPath;
-        private final Handler.Callback mOriginalCallback;
-
-        private final int LAUNCH_ACTIVITY;
-        private final int RELAUNCH_ACTIVITY;
-        private final int EXECUTE_TRANSACTION;
-
-        private Method mGetCallbacksMethod = null;
-        private boolean mSkipInterceptExecuteTransaction = false;
-
-        ResourceInsuranceHandlerCallback(Context context, String patchResApkPath, Handler.Callback original, Class<?> hClazz) {
-            Context appContext = context.getApplicationContext();
-            mContext = (appContext != null ? appContext : context);
-            mPatchResApkPath = patchResApkPath;
-            mOriginalCallback = original;
-            LAUNCH_ACTIVITY = fetchMessageId(hClazz, "LAUNCH_ACTIVITY", 100);
-            RELAUNCH_ACTIVITY = fetchMessageId(hClazz, "RELAUNCH_ACTIVITY", 126);
-
-            if (ShareTinkerInternals.isNewerOrEqualThanVersion(28, true)) {
-                EXECUTE_TRANSACTION  = fetchMessageId(hClazz, "EXECUTE_TRANSACTION ", 159);
-            } else {
-                EXECUTE_TRANSACTION = -1;
-            }
-        }
-
-        private int fetchMessageId(Class<?> hClazz, String name, int defVal) {
-            int value;
-            try {
-                value = ShareReflectUtil.findField(hClazz, name).getInt(null);
-            } catch (Throwable e) {
-                value = defVal;
-            }
-            return value;
-        }
-
-        @Override
-        public boolean handleMessage(Message msg) {
-            boolean consume = false;
-            if (hackMessage(msg)) {
-                consume = true;
-            } else if (mOriginalCallback != null) {
-                consume = mOriginalCallback.handleMessage(msg);
-            }
-            return consume;
-        }
-
-        @SuppressWarnings("unchecked")
-        private boolean hackMessage(Message msg) {
-            boolean shouldReInjectPatchedResources = false;
-            if (!isPatchedResModifiedAfterLastLoad(mPatchResApkPath)) {
-                shouldReInjectPatchedResources = false;
-            } else {
-                if (msg.what == LAUNCH_ACTIVITY || msg.what == RELAUNCH_ACTIVITY) {
-                    shouldReInjectPatchedResources = true;
-                } else if (msg.what == EXECUTE_TRANSACTION) {
-                    do {
-                        if (mSkipInterceptExecuteTransaction) {
-                            break;
-                        }
-                        final Object transaction = msg.obj;
-                        if (transaction == null) {
-                            ShareTinkerLog.w(TAG, "transaction is null, skip rest insurance logic.");
-                            break;
-                        }
-                        if (mGetCallbacksMethod == null) {
-                            try {
-                                mGetCallbacksMethod = ShareReflectUtil.findMethod(transaction, "getCallbacks");
-                            } catch (Throwable ignored) {
-                                // Ignored.
-                            }
-                        }
-                        if (mGetCallbacksMethod == null) {
-                            ShareTinkerLog.e(TAG, "fail to find getLifecycleStateRequest method, skip rest insurance logic.");
-                            mSkipInterceptExecuteTransaction = true;
-                            break;
-                        }
-                        try {
-                            final List<Object> req = (List<Object>) mGetCallbacksMethod.invoke(transaction);
-                            if (req != null && req.size() > 0) {
-                                final Object cb = req.get(0);
-                                shouldReInjectPatchedResources = cb != null && cb.getClass().getName().equals(LAUNCH_ACTIVITY_LIFECYCLE_ITEM_CLASSNAME);
-                            }
-                        } catch (Throwable ignored) {
-                            ShareTinkerLog.e(TAG, "fail to call getLifecycleStateRequest method, skip rest insurance logic.");
-                        }
-                    } while (false);
-                }
-            }
-            if (shouldReInjectPatchedResources) {
-                try {
-                    monkeyPatchExistingResources(mContext, mPatchResApkPath, true);
-                } catch (Throwable thr) {
-                    ShareTinkerLog.printErrStackTrace(TAG, thr, "fail to ensure patched resources available after it's modified.");
-                }
-            }
-            return false;
-        }
-    }
-
-    private static boolean isPatchedResModifiedAfterLastLoad(String patchedResPath) {
-        long patchedResModifiedTime;
-        try {
-            patchedResModifiedTime = new File(patchedResPath).lastModified();
-        } catch (Throwable thr) {
-            ShareTinkerLog.printErrStackTrace(TAG, thr, "Fail to get patched res modified time.");
-            patchedResModifiedTime = 0L;
-        }
-        if (patchedResModifiedTime == 0) {
-            return false;
-        }
-        if (patchedResModifiedTime == storedPatchedResModifiedTime) {
-            return false;
-        }
-        return true;
-    }
-
-    private static void recordCurrentPatchedResModifiedTime(String patchedResPath) {
-        try {
-            storedPatchedResModifiedTime = new File(patchedResPath).lastModified();
-        } catch (Throwable thr) {
-            ShareTinkerLog.printErrStackTrace(TAG, thr, "Fail to store patched res modified time.");
-            storedPatchedResModifiedTime = 0L;
+        if (!PatchedResourcesInsuranceLogic.install(context, externalResourceFile)) {
+            ShareTinkerLog.w(TAG, "tryLoadPatchFiles:PatchedResourcesInsuranceLogic install fail.");
+            throw new TinkerRuntimeException("fail to install PatchedResourcesInsuranceLogic.");
         }
     }
